% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{crmatrix}
\alias{crmatrix}
\title{Shorthand to create a matrix with \code{spnum} rows and \code{resnum} columns}
\usage{
crmatrix(..., byrow = TRUE)
}
\arguments{
\item{...}{One or more numeric vectors that, concatenated by \code{c()}, produce a vector of length \code{spnum * resnum}.}

\item{byrow}{A logical vector of length 1, equivalent to \code{byrow} of \code{matrix()}.}
}
\value{
S3 object of class \code{rescomp_crmatrix}.
}
\description{
This produces an object that, when passed through \code{rescomp} functions such as \code{funcresp_type1()}, and ultimately provided to \code{spec_rescomp()}, inherits \code{spnum} and \code{resnum} from the arguments to \code{spec_rescomp()}.
Thus, it is shorthand for \code{matrix(c(...), nrow = spnum, ncol = resnum, byrow = TRUE)}.
}
\examples{
# TODO
}
