% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{def_cr_ode}
\alias{def_cr_ode}
\title{Define consumer resource ODE function}
\usage{
def_cr_ode(t, y, pars)
}
\arguments{
\item{t}{The current time of the simulation.}

\item{y}{The vector of current estimates of consumers and resources in the simulation.}

\item{pars}{S3 object of class \code{rescomp} containing model specification.}
}
\value{
List of length one, containing a vector of derivatives of \code{y} with respect to time.
}
\description{
Define consumer resource ODE function
}
