% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_stochastic.R
\name{def_cr_transition_rates}
\alias{def_cr_transition_rates}
\title{Get the rates of transitions for a Gillespie algorithm simulation}
\usage{
def_cr_transition_rates(y, pars, t)
}
\arguments{
\item{y}{The vector of current estimates of consumers and resources in the simulation.}

\item{pars}{The current time of the simulation.}
}
\value{
A vector of rates of each of the transitions produced by \code{def_cr_transitions()}.
}
\description{
Get the rates of transitions for a Gillespie algorithm simulation
}
\examples{
# TODO
}
