% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_stochastic.R
\name{def_cr_transitions}
\alias{def_cr_transitions}
\title{Get the list of transitions for a Gillespie algorithm simulation}
\usage{
def_cr_transitions(pars)
}
\arguments{
\item{pars}{The \code{rescomp} object passed to \code{sim_rescomp()}.}
}
\value{
A list of transitions suitable to pass to \code{adaptivetau::ssa.adaptivetau()}.
}
\description{
Get the list of transitions for a Gillespie algorithm simulation
}
\examples{
# TODO
}
