% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_funcresp.R
\name{df_funcresp}
\alias{df_funcresp}
\title{Generate ggplot friendly dataframe for plotting functional responses}
\usage{
df_funcresp(pars, maxx)
}
\arguments{
\item{pars}{parameter list from make_par_list()}

\item{maxx}{maximum resource value to get percapita growth rates
(for plotting)}
}
\value{
data frame
}
\description{
Generate ggplot friendly dataframe for plotting functional responses
}
