% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_batch_transfer}
\alias{event_batch_transfer}
\title{Define an event in which the species and resources are diluted by addition of new medium}
\usage{
event_batch_transfer(dilution, resources)
}
\arguments{
\item{dilution}{A numeric vector or \code{rescomp_coefs_vector} of length 1 (whose value should be
between 0 and 1) representing the proportion of the original medium retained.}

\item{resources}{A numeric vector or \code{rescomp_coefs_vector} of resource concentrations.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object representing a batch transfer by dilution with new medium.
First, all existing species and resource concentrations are multiplied by \code{dilution}.
Then, the resource concentrations listed in \code{resources}, multiplied by 1 - \code{dilution}, are added.
}
\examples{
# TODO
}
