% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_res_add}
\alias{event_res_add}
\title{Define an event in which a constant amount of some resources is added to (or removed from) the system}
\usage{
event_res_add(resources, min_zero = TRUE)
}
\arguments{
\item{resources}{A numeric vector or \code{rescomp_coefs_vector} of resource concentrations, by which the current resource concentrations are increased. Can be negative, to decrease resource concentrations.}

\item{min_zero}{If this is TRUE, resulting resource concentrations are clamped to a minimum of zero. If this is FALSE, negative values of \code{resources} may reduce resource concentrations below zero.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object representing a pulse of added/removed resources.
}
\examples{
m1 <- spec_rescomp(
  spnum = 2, resnum = 2,
  funcresp = funcresp_type1(crmatrix(
    0.10, 0.08,
    0.06, 0.12
  )),
  events = list(
    event_schedule_periodic(
      event_res_add(c(-0.1, 0.1)),
      period = 250
    )
  )
)
plot_rescomp(sim_rescomp(m1))
}
