% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_res_mult}
\alias{event_res_mult}
\title{Define an event in which the resource populations are multiplied by some factor}
\usage{
event_res_mult(resources_mult)
}
\arguments{
\item{resources_mult}{A numeric vector or \code{rescomp_coefs_vector} by which to multiply the current resource concentrations. Should be greater than 1 for an increase, or less than 1 for a decrease. Should not be negative.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object a pulse of multiplicatively increased/decreased resource populations, e.g. a density-independent disturbance which kills prey species.
}
\examples{
m1 <- spec_rescomp(
  spnum = 2, resnum = 2,
  funcresp = funcresp_type1(crmatrix(
    0.10, 0.08,
    0.06, 0.12
  )),
  events = list(
    event_schedule_periodic(
      event_res_mult(c(0.5, 1.5)),
      period = 250
    )
  )
)
plot_rescomp(sim_rescomp(m1))
}
