% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_res_pulse}
\alias{event_res_pulse}
\title{Define an event in which a constant amount of some resources is added to the system}
\usage{
event_res_pulse(resources)
}
\arguments{
\item{resources}{A numeric vector or \code{rescomp_coefs_vector} of resource concentrations, by which the current resource concentrations are increased.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object representing a resource pulse.
}
\examples{
# TODO
}
