% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_schedule.R
\name{event_schedule_fixed}
\alias{event_schedule_fixed}
\title{Define a schedule of an event occurring at a fixed set of times}
\usage{
event_schedule_fixed(event_obj, times, priority = 0)
}
\arguments{
\item{event_obj}{An object of class \code{rescomp_event}.}

\item{times}{A numeric vector of times at which the event should occur.}

\item{priority}{A number. If two events occur at the same time, the event with the lower priority number is processed first.}
}
\value{
S3 object of class \code{rescomp_event_schedule}.
}
\description{
Produces an event schedule object suitable for passing in the \code{events} argument to \code{spec_rescomp()}.
}
\examples{
m1 <- spec_rescomp(
  events = list(
    event_schedule_fixed(
      event_res_add(0.5),
      times = c(200, 400)
    ),
    event_schedule_fixed(
      event_sp_add(-500),
      times = c(600, 800)
    )
  )
)
plot_rescomp(sim_rescomp(m1))
}
