% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_schedule.R
\name{event_schedule_periodic}
\alias{event_schedule_periodic}
\title{Define a schedule of an event occurring repeatedly at a fixed period}
\usage{
event_schedule_periodic(event_obj, period, start_time = period, priority = 0)
}
\arguments{
\item{event_obj}{An object of class \code{rescomp_event}.}

\item{period}{A number: the period at which the event occurs.}

\item{start_time}{A number: the time at which the event first occurs.}

\item{priority}{A number. If two events occur at the same time, the event with the lower priority number is processed first.}
}
\value{
S3 object of class \code{rescomp_event_schedule}.
}
\description{
Produces an event schedule object suitable for passing in the \code{events} argument to \code{spec_rescomp()}.
}
\examples{
# TODO
}
