% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_set.R
\name{event_set_introseq}
\alias{event_set_introseq}
\title{Define a schedule of one-off species introductions: one for each species.}
\usage{
event_set_introseq(times, concentrations, priority = 0)
}
\arguments{
\item{times}{A numeric vector, of length \code{spnum}, of times at which each species should be introduced.}

\item{concentrations}{A numeric vector or \code{rescomp_coefs_vector}, of length \code{spnum}, of species concentrations; the concentration of each species that should be added at the corresponding time.}

\item{priority}{A number. If two events occur at the same time, the event with the lower priority number is processed first.}
}
\value{
S3 object of class \code{rescomp_event_schedule}.
}
\description{
Produces an event schedule object suitable for passing in the \code{events} argument to \code{spec_rescomp()}.
}
\examples{
# TODO
}
