% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_sp_add}
\alias{event_sp_add}
\title{Define an event in which a constant amount of some consumers is added to (or removed from) the system}
\usage{
event_sp_add(species, min_zero = TRUE)
}
\arguments{
\item{species}{A numeric vector or \code{rescomp_coefs_vector} of species concentrations, by which the current species concentrations are increased. Can be negative, to decrease species concentrations.}

\item{min_zero}{If this is TRUE, resulting species concentrations are clamped to a minimum of zero. If this is FALSE, negative values of \code{species} may reduce species concentrations below zero.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object representing a pulse of added/removed consumers, e.g. a species introduction.
}
\examples{
# TODO
}
