% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_sp_custom}
\alias{event_sp_custom}
\title{Define an event which instantaneously changes species concentrations using an arbitrary function}
\usage{
event_sp_custom(func, spnum = NULL)
}
\arguments{
\item{func}{A function that takes \code{species} (a numeric vector of species concentrations) and \code{params} (a list of parameters) and returns a numeric vector of species concentrations.}

\item{spnum}{The number of species.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object suitable for building event schedules to pass \code{spec_rescomp()}.
}
\details{
If \code{spnum} is NULL, \code{spec_rescomp()} will attempt to infer it.
This is fine if the result is used directly in an event schedule, but may fail if the result must be combined with other \code{rescomp_event}s first.
}
\examples{
pars <- spec_rescomp(
  events = list(
    event_schedule_fixed(
      event = event_sp_custom(function(sp, params) {
        sp * 2
      }),
      times = 500
    )
  )
)
plot_rescomp(sim_rescomp(pars))
}
