% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{eventfun_respulse}
\alias{eventfun_respulse}
\title{Event for resource pulsing}
\usage{
eventfun_respulse(Time, State, Pars)
}
\arguments{
\item{Time}{Time to simulate over}

\item{State}{Vector of initial states}

\item{Pars}{List}
}
\description{
Event for resource pulsing
}
