% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{eventfun_starttime}
\alias{eventfun_starttime}
\title{Event for different consumer start times}
\usage{
eventfun_starttime(Time, State, Pars)
}
\arguments{
\item{Time}{Time to simulate over}

\item{State}{Vector of initial states}

\item{Pars}{List}
}
\description{
Event for different consumer start times
}
