% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_and_name.R
\name{frame_and_name}
\alias{frame_and_name}
\title{Convert object of class deSolve to a data frame and name columns}
\usage{
frame_and_name(odeobject, pars)
}
\arguments{
\item{odeobject}{object of class deSolve}

\item{pars}{Model parameters from make_par_list()}
}
\value{
data frame
}
\description{
Convert object of class deSolve to a data frame and name columns
}
\examples{
library(deSolve)
pars <- make_par_list()
happenings <- time_vals()
m1 <- ode(
    func = def_cr_ode,
    y = initiate_state(vars = pars),
    parms = pars,
    times = happenings$totaltime,
    method = "lsoda"
)
frame_and_name(m1, pars)

}
