% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_form.R
\name{func_form}
\alias{func_form}
\title{Generate type 1,2 or 3 functional response}
\usage{
func_form(R, mu, Ks, phi, type3)
}
\arguments{
\item{R}{vector of resource concentrations}

\item{mu}{maximum growth rate}

\item{Ks}{half saturation constant}

\item{phi}{0 if linear, 1 otherwise}

\item{type3}{1 if type 3, 1/2 otherwise}
}
\value{

}
\description{
Generate type 1,2 or 3 functional response
}
