% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_hill}
\alias{funcresp_hill}
\title{Define a Hill functional response}
\usage{
funcresp_hill(mumax, ks, n)
}
\arguments{
\item{mumax}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The maximum growth rate (if using quota) or maximum consumption rate (if using efficiency) of each species on each resource.}

\item{ks}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The half saturation constant of each species for each resource; the resource concentration at which growth/consumption rate is half of mumax.}

\item{n}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The Hill coefficient for each species on each resource. With n = 1, this is equivalent to a Monod functional response.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a Hill functional response with maximum growth rate \code{mumax}, half saturation constant \code{ks}, and Hill coefficient n.
mu_ij(R_j) = mumax_ij * (R_j)^n_ij / (ks_ij + (R_j)^n_ij)
This an alternative parameterisation of a type 3 functional response, typically used with quota rather than efficiency.
}
\examples{
# TODO
}
