% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_monod}
\alias{funcresp_monod}
\title{Define a Monod functional response}
\usage{
funcresp_monod(mumax, ks)
}
\arguments{
\item{mumax}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The maximum growth rate (if using quota) or maximum consumption rate (if using efficiency) of each species on each resource.}

\item{ks}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The half saturation constant of each species for each resource; the resource concentration at which growth/consumption rate is half of mumax.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a Monod functional response with maximum growth rate \code{mumax} and half saturation constant \code{ks}.
mu_ij(R_j) = mumax_ij * R_j / (ks_ij + R_j)
This an alternative parameterisation of a type 2 functional response, typically used with quota rather than efficiency.
}
\examples{
# TODO
}
