% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_type1}
\alias{funcresp_type1}
\title{Define a linear functional response}
\usage{
funcresp_type1(a)
}
\arguments{
\item{a}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The attack rate (if using efficiency) or growth rate (if using quota) of each species on each resource.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a linear or type 1 functional response with attack rate \code{a}.
mu_ij(R_j) = a_ij * R_j
}
\examples{
m1 <- spec_rescomp(
  spnum = 2, resnum = 2,
  funcresp = funcresp_type1(
    crmatrix(0.2, 0.3, 0.4, 0.2)
  )
)
plot_funcresp(m1)

m2 <- spec_rescomp(
  spnum = 2, resnum = 2,
  funcresp = funcresp_type1(
    rescomp_coefs_lerp(
      crmatrix(0.2, 0.3, 0.4, 0.2),
      crmatrix(0.2, 0.3, 0.1, 0.1),
      "antibiotic_concentration"
    )
  ),
  params = list(antibiotic_concentration = rescomp_param_square())
)
plot_funcresp(m2, display_values = list(antibiotic_concentration = c(0, 0.5, 1)))
}
