% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_type2}
\alias{funcresp_type2}
\title{Define a Holling type 2 functional response}
\usage{
funcresp_type2(a, h)
}
\arguments{
\item{a}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The maximum attack rate (if using efficiency) or growth rate (if using quota) of each species on each resource, when the resource is rare.}

\item{h}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The handling time of each species for each resource, which reduces attack/growth rate when the resource becomes abundant. With h = 0, this is equivalent to a type 1 function response.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a Holling type 2 functional response with attack rate \code{a} and handling time \code{h}.
mu_ij(R_j) = (a_ij * R_j) / (1 + a_ij * h_ij * R_j)
}
\examples{
# TODO
}
