% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_type3}
\alias{funcresp_type3}
\title{Define a Holling type 3 functional response}
\usage{
funcresp_type3(a, h, k)
}
\arguments{
\item{a}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The maximum attack rate (if using efficiency) or growth rate (if using quota) of each species on each resource, when the resource is rare.}

\item{h}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The handling time of each species for each resource, which reduces attack/growth rate when the resource becomes abundant.}

\item{k}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The exponent (a.k.a. Hill coefficient) for each species on each resource. With k = 1, this is equivalent to a type 2 functional response. Each k must be greater than 1 for a true type 3 functional response.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a Holling type 3 functional response with attack rate \code{a}, handling time \code{h}, and exponent \code{k}.
mu_ij(R_j) = (a_ij * (R_j)^k_ij) / (1 + a_ij * h_ij * (R_j)^k_ij)
}
\examples{
# TODO
}
