% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{get_coefs}
\alias{get_coefs}
\alias{get_coefs_vector}
\alias{get_coefs_matrix}
\title{Get coefficients from a vector/matrix or \code{rescomp_coefs_vector}/\code{rescomp_coefs_matrix} object}
\usage{
get_coefs(coefs_obj, params)

get_coefs_vector(coefs_obj, params)

get_coefs_matrix(coefs_obj, params)
}
\arguments{
\item{coefs_obj}{A numeric vector/matrix or an object of class \code{rescomp_coefs_vector}/\code{rescomp_coefs_matrix}.}

\item{params}{A list of time-dependent parameters.}
}
\value{
A vector/matrix of coefficients.
}
\description{
Provides a generic interface to a vector/matrix or a \code{rescomp_coefs_vector}/\code{rescomp_coefs_matrix} object.
Called on a numeric vector/matrix (as appropriate to the function used), this will just return the vector/matrix.
Called on a \code{rescomp_coefs_vector}/\code{rescomp_coefs_matrix}, this allows time-dependence via \code{parameters}.
\code{get_coefs()} works on vector- or matrix-type objects, while the others work only on their specific types.
}
\details{
This function is normally only for internal use, but is exported to aid in debugging created \code{rescomp_coefs} objects.
}
\examples{
get_coefs(c(0.2, 0.3, 0.4), list())
get_coefs(matrix(c(0.2, 0.3, 0.4, 0.2), nrow = 2), list())
get_coefs_vector(c(0.2, 0.3, 0.4), list())
try(get_coefs_matrix(c(0.2, 0.3, 0.4), list()))

coefs <- rescomp_coefs_matrix_custom(
  function(params) {
    matrix(c(0.2, 0.3, 0.4, params$fourth_coeff), nrow = 2, ncol = 2)
  },
  nrow = 2, ncol = 2
)
get_coefs(coefs, list(fourth_coeff = 0.5))
}
