% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_schedule.R
\name{get_event_times}
\alias{get_event_times}
\title{Gets the list of times from a \code{rescomp_event_schedule} object}
\usage{
get_event_times(event_schedule_obj, totaltime)
}
\arguments{
\item{event_schedule_obj}{An object of class \code{rescomp_event_schedule}.}

\item{totaltime}{The total time which the simulation will run for.}
}
\value{
A vector of times.
}
\description{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_event_schedule} objects.
}
\examples{
schedule1 <- event_schedule_periodic(
  event_batch_transfer(dilution = 0.1, resources = 1),
  period = 250
)
get_event_times(schedule1, 1000)
get_event_times(schedule1, 999)

schedule2 <- event_schedule_periodic(
  event_batch_transfer(dilution = 0.1, resources = 1),
  period = 250,
  start_time = 125
)
get_event_times(schedule2, 1000)
get_event_times(schedule2, 999)
}
