% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{get_funcresp}
\alias{get_funcresp}
\title{Get growth rates from a \code{rescomp_funcresp} object}
\usage{
get_funcresp(funcresp_obj, spnum, resources, params)
}
\arguments{
\item{funcresp_obj}{An object of class \code{rescomp_funcresp}.}

\item{spnum}{The number of species.}

\item{resources}{A vector of resource concentrations.}

\item{params}{A list of time-dependent parameters.}
}
\value{
A matrix of species growth rates on each resource, with \code{spnum} rows and \code{length(resources)} columns.
}
\description{
Gets the growth rates of each species on each resource, given resource concentrations.
These must be combined, according to whether the resouces are essential or substitutable, to get the overall growth rate for each species.
}
\details{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_funcresp} objects.
}
\examples{
funcresp <- spec_funcresp_custom(
  function(resources, params) {
    growth_rates <- params$scale * c(0.2, 0.3)
    outer(growth_rates, resources)
  },
  spnum = 2
)
get_funcresp(funcresp, 2, c(1, 4, 5, 6), list(scale = 2))
get_funcresp(funcresp, 2, 0.7, list(scale = 0.5))
try(get_funcresp(funcresp, 3, 0.7, list(scale = 0.5)))
}
