% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_params}
\alias{get_params}
\title{Get params at an instant in time from a \code{rescomp_param} object}
\usage{
get_params(param_obj, t)
}
\arguments{
\item{param_obj}{An object of class \code{rescomp_param}.}

\item{t}{The time at which to get the parameters.}
}
\value{
The parameters; various objects depending on the type of \code{param_obj}.
}
\description{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_param} objects.
}
\examples{
# Repeated pulsing and exponential decay
antibiotic_conc <- rescomp_param_custom(function(t) {
  0.5^(10 * (t \%\% 1))
})
get_params(antibiotic_conc, 0)
get_params(antibiotic_conc, 0.5)
get_params(antibiotic_conc, 1)

params <- rescomp_param_list(r = 0.2, antibiotic_conc = antibiotic_conc)
get_params(params, 0.5)
}
