% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ressupply.R
\name{get_ressupply}
\alias{get_ressupply}
\alias{get_ressupply_increase}
\alias{get_ressupply_decrease}
\title{Get resource supply rates from a \code{rescomp_ressupply} object}
\usage{
get_ressupply(ressupply_obj, resources, params)

get_ressupply_increase(ressupply_obj, resources, params)

get_ressupply_decrease(ressupply_obj, resources, params)
}
\arguments{
\item{ressupply_obj}{An object of class \code{rescomp_funcresp}.}

\item{resources}{A vector of resource concentrations.}

\item{params}{A list of time-dependent parameters.}
}
\value{
A vector of rates of change of resource concentrations, of the same length as \code{resources}.
}
\description{
\code{get_ressupply} gets the resource supply rates of each resource, given the current resource concentrations.
It is simply the difference between the results of \code{get_ressupply_increase} and ``get_ressupply_decrease`.
It is only important that increases and decreases are treated differently when performing stochastic simulations,
as these are the independent rates of positive and negative transitions.
}
\details{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_ressupply} objects.
}
\examples{
# Two resources, A and B, with constant supply of A, and A spontaneously converting to B
ressupply <- ressupply_custom(
  function(resources, params) {
    conversion <- params$conversion * resources[1]
    return(c(params$supply - conversion, conversion))
  },
  resnum = 2
)
get_ressupply(ressupply, c(10, 20), list(supply = 3, conversion = 0.2))
try(get_ressupply(ressupply, c(10, 20, 30), list(supply = 3, conversion = 0.2)))
}
