% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_state.R
\name{initiate_state}
\alias{initiate_state}
\title{Providing starting values for state variables}
\usage{
initiate_state(pars)
}
\arguments{
\item{pars}{Model parameters from make_par_list()}
}
\value{
vector
}
\description{
Providing starting values for state variables
}
\examples{
initiate_state(pars = make_par_list())
}
