% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_sim.R
\name{lin_sim}
\alias{lin_sim}
\title{Generate a linear (type 1) functional response}
\usage{
lin_sim(x, r)
}
\arguments{
\item{x}{vector of resource concentrations}

\item{r}{constant of proportionality}
}
\value{
vector of per capita growth rates
}
\description{
Generate a linear (type 1) functional response
}
\examples{
lin_sim(seq(1,10,0.1), r = 1)
}
