% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_list.R
\name{make_par_list}
\alias{make_par_list}
\title{Generate list of parameters for consumer-resource model}
\usage{
make_par_list(
  spnum = 2,
  resnum = 1,
  mumatrix,
  kmatrix,
  qmatrix,
  linear = TRUE,
  essential = FALSE,
  mort = 0.03,
  chemospeed = 10,
  chemoconc = 1,
  logisK,
  logisr
)
}
\arguments{
\item{spnum}{number of consumers}

\item{resnum}{number of resources}

\item{mumatrix}{matrix of maximum growth rates (type 2)/ constant of proportionalities (type 1). # rows = spnum; # cols = resnum}

\item{kmatrix}{matrix of half saturation constants (type 2, ignored if linear = TRUE). # rows = spnum; # cols = resnum}

\item{qmatrix}{matrix of resource quotas. # rows = spnum; # cols = resnum}

\item{linear}{if FALSE equals type 2 function (i.e. Monod)}

\item{essential}{if FALSE resources are substitutable}

\item{mort}{density independent mortality rate}

\item{chemospeed}{chemostat dilution rate}

\item{chemoconc}{chemostat supply concentration}

\item{logisK}{resource carrying capcity (if chemo = FALSE)}

\item{logisr}{resource intinsic rate of increase (if chemo = FALSE)}
}
\value{

}
\description{
Generate list of parameters for consumer-resource model
}
\examples{
make_par_list()
}
