% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tdpars.R
\name{make_tdpars}
\alias{make_tdpars}
\title{Internal function for making time dependent parameters}
\usage{
make_tdpars(crpars, pars)
}
\arguments{
\item{crpars}{Focal parameter}

\item{pars}{Immature pars list}
}
\value{
A list of approximation functions
}
\description{
Internal function for making time dependent parameters
}
