% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monod_sim.R
\name{monod_sim}
\alias{monod_sim}
\title{Generate a Monod (type 2) functional response}
\usage{
monod_sim(x, r, k)
}
\arguments{
\item{x}{vector of resource concentrations}

\item{r}{numeric}

\item{k}{numeric}
}
\value{
vector
}
\description{
Generate a Monod (type 2) functional response
}
\examples{
monod_sim(seq(1,10,0.1), r = 1, k = 0.1)
}
