% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crsim.R
\name{plot_crsim}
\alias{plot_crsim}
\title{Plot consumer and resource dynamics from deSolve output}
\usage{
plot_crsim(odeobj, pars)
}
\arguments{
\item{odeobj}{odeobject}

\item{pars}{parameters from spec_rescomp()}
}
\value{
ggplot object
}
\description{
Plot consumer and resource dynamics from deSolve output
}
\examples{

pars <- spec_rescomp()
m1 <- sim_rescomp(pars)
plot_crsim(m1, pars)

}
