% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcresp.R
\name{plot_funcresp}
\alias{plot_funcresp}
\title{Plot functional responses}
\usage{
plot_funcresp(pars, maxx)
}
\arguments{
\item{pars}{Parameter list from make_par_list()}

\item{maxx}{Resource value to calculate per-capita growth rates up to (xlim).}
}
\value{
ggplot object
}
\description{
Plot functional responses
}
\examples{

pars <- make_par_list()
plot_funcresp(pars)

pars <- make_par_list(
    spnum = 2,
    resnum = 2,
    linear = FALSE,
    mumatrix = list(matrix(c(0.7,0.3,
                             0.4,0.5),
                           nrow = 2,
                           ncol = 2,
                           byrow = TRUE))
)
plot_funcresp(pars)

}
