% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rescomp.R
\name{plot_rescomp}
\alias{plot_rescomp}
\title{Plot consumer and resource dynamics from deSolve output}
\usage{
plot_rescomp(model, consumers = TRUE, resources = TRUE)
}
\arguments{
\item{model}{List output from \code{sim_rescomp()}. First element is an object of
class deSolve. Second element is an object of class rescomp.}

\item{consumers}{Plot consumer dynamics? Default = TRUE.}

\item{resources}{Plot resource dynamics? Default = TRUE.}
}
\value{
ggplot object
}
\description{
Plot consumer and resource dynamics from deSolve output
}
\examples{

pars <- spec_rescomp()
m1 <- sim_rescomp(pars)
plot_rescomp(m1)

}
