% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_rescomp.R
\name{print.rescomp}
\alias{print.rescomp}
\title{Print method for class rescomp}
\usage{
\method{print}{rescomp}(pars, detail = "summary")
}
\arguments{
\item{pars}{rescomp object}

\item{detail}{Either "summary" or "list". "summary" (default) provides plain language
summary info. "list" prints the full parameter list.}
}
\value{

}
\description{
Print method for class rescomp
}
