% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{rescomp_coefs_lerp}
\alias{rescomp_coefs_lerp}
\title{Create a set of coefficients by linear interpolation}
\usage{
rescomp_coefs_lerp(coefs0, coefs1, param_name, param0 = 0, param1 = 1)
}
\arguments{
\item{coefs0, coefs1}{Vectors, matrices, or objects of class \code{rescomp_coefs_vector} or \code{rescomp_coefs_matrix} to be interpolated between. Either both must be vectors, or both must be matrices.}

\item{param_name}{A character vector of length 1; the name of the parameter to be interpolated according to.}

\item{param0, param1}{The values of the parameter to take as the fixed points for interpolation.}
}
\value{
S3 object of class \code{rescomp_coefs_vector} or \code{rescomp_coefs_matrix}, according to the types of \code{coefs0} and \code{coefs1}.
}
\description{
Produces an object that may be used in place of a vector or matrix in the creation of arguments to \code{spec_rescomp()}.
Linearly interpolates between two sets of coefficients according to the value of a time-dependent parameter.
}
\details{
Coefficients are taken to have the values in \code{coef0} at \code{param = param0}, and the values in \code{coef1} at \code{param = param1}, and are linearly interpolated between.
Each parameter in the vector/matrix is interpolated independently.
}
\examples{
coefs_vec <- rescomp_coefs_lerp(
  c(0, 0, 0),
  c(2, 3, 5),
  "scale"
)
get_coefs(coefs_vec, list(scale = 0))
get_coefs(coefs_vec, list(scale = 0.5))
get_coefs(coefs_vec, list(scale = 1))
get_coefs(coefs_vec, list(scale = 1.5))

coefs_mat1 <- rescomp_coefs_lerp(
  matrix(c(0.2, 0.4, 0.3, 0.2), nrow = 2),
  matrix(c(0.2, 0.1, 0.3, 0.1), nrow = 2),
  "antibiotic1_concentration",
  param0 = 0,
  param1 = 200
)
get_coefs(coefs_mat1, list(antibiotic1_concentration = 0))
get_coefs(coefs_mat1, list(antibiotic1_concentration = 100))
get_coefs(coefs_mat1, list(antibiotic1_concentration = 200))

coefs_mat2 <- rescomp_coefs_lerp(
  coefs_mat1,
  matrix(c(0.0, 0.0, 0.0, 0.0), nrow = 2),
  "antibiotic2_concentration",
  param0 = 0,
  param1 = 100
)
get_coefs(coefs_mat2, list(antibiotic1_concentration = 100, antibiotic2_concentration = 50))
}
