% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{rescomp_coefs_matrix_custom}
\alias{rescomp_coefs_matrix_custom}
\title{Create a matrix of coefficients using an arbitrary function}
\usage{
rescomp_coefs_matrix_custom(func, nrow, ncol)
}
\arguments{
\item{func}{A function that takes \code{params} (a list of parameters) and returns a matrix.}

\item{nrow}{The number of rows in the matrix returned by \code{func}.}

\item{ncol}{The number of columns in the matrix returned by \code{func}.}
}
\value{
S3 object of class \code{rescomp_coefs_matrix}.
}
\description{
Produces an object that may be used in place of a matrix in the creation of arguments to \code{spec_rescomp()}.
}
\examples{
# Matrix with one time-varying coefficient
rescomp_coefs_matrix_custom(
  function(params) {
    matrix(c(0.2, 0.3, 0.4, params$fourth_coeff), nrow = 2, ncol = 2)
  },
  nrow = 2, ncol = 2
)
}
