% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefs.R
\name{rescomp_coefs_vector_custom}
\alias{rescomp_coefs_vector_custom}
\title{Create a vector of coefficients using an arbitrary function}
\usage{
rescomp_coefs_vector_custom(func, length)
}
\arguments{
\item{func}{A function that takes \code{params} (a list of parameters) and returns a vector.}

\item{length}{The length of the vector returned by \code{func}.}
}
\value{
S3 object of class \code{rescomp_coefs_vector}.
}
\description{
Produces an object that may be used in place of a vector in the creation of arguments to \code{spec_rescomp()}.
}
\examples{
# Vector with one time-varying coefficient
rescomp_coefs_vector_custom(
  function(params) {
    c(0.2, 0.3, 0.4, params$fourth_coeff)
  },
  length = 4
)
}
