% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{rescomp_param_list}
\alias{rescomp_param_list}
\title{Create a list of named rescomp parameters}
\usage{
rescomp_param_list(...)
}
\arguments{
\item{...}{Named arguments used to create the params list.}
}
\value{
S3 object of class \code{rescomp_param_list}.
}
\description{
Produces an object suitable to pass as the \code{params} to \code{spec_rescomp}.
}
\details{
Must be given zero or more named arguments, which are used to build the params list.
\code{rescomp_param} objects are parsed to allow time dependence, while other objects are passed into the list directly.
}
\examples{
params <- rescomp_param_list(
  r = 0.2,
  s = c(3, 4),
  conc = rescomp_param_custom(function(t) {
    t^2
  })
)
get_params(params, 0.5)
get_params(params, 2)
}
