% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ressupply.R
\name{ressupply_chemostat}
\alias{ressupply_chemostat}
\title{Create a resource supply rate using chemostat dynamics}
\usage{
ressupply_chemostat(dilution, concentration)
}
\arguments{
\item{dilution}{A numeric vector or \code{rescomp_coefs_vector}, of length one.}

\item{concentration}{A vector or \code{rescomp_coefs_vector}, with one number per resource. The concentration of each resource in the incoming medium.}
}
\value{
S3 object of class \code{rescomp_ressupply}.
}
\description{
Produces an object suitable to pass as the \code{ressupply} to \code{spec_rescomp()}.
}
\examples{
ressupply <- ressupply_chemostat(
  dilution = 0.01,
  concentration = rescomp_coefs_lerp(c(0, 0, 0), c(2, 3, 4), "ressupply_scaling")
)
get_ressupply(ressupply, c(2, 4, 10), list(ressupply_scaling = 0))
get_ressupply(ressupply, c(3, 3, 3), list(ressupply_scaling = 1))
}
