% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ressupply.R
\name{ressupply_constant}
\alias{ressupply_constant}
\title{Create a resource supply rate using a constant rate resource supply}
\usage{
ressupply_constant(rate)
}
\arguments{
\item{rate}{A vector or \code{rescomp_coefs_vector}, with one number per resource. The supply rate of each resource.}
}
\value{
S3 object of class \code{rescomp_ressupply}.
}
\description{
Produces an object suitable to pass as the \code{ressupply} to \code{spec_rescomp()}.
}
\examples{
ressupply <- ressupply_constant(c(0.2, 0.3))
get_ressupply(ressupply, c(2, 10), list())
get_ressupply(ressupply, c(5, 20), list())
# The above two give the same result; constant supply doesn't depend on existing concentration.

ressupply <- ressupply_constant(rescomp_coefs_lerp(c(0.2, 0.3), c(0.4, 0.6), "extra_supply"))
get_ressupply(ressupply, c(2, 10), list(extra_supply = 0.2))
get_ressupply(ressupply, c(2, 10), list(extra_supply = 0.8))
}
