% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rescomp.R
\name{sim_rescomp}
\alias{sim_rescomp}
\title{Simulate resource competition (a convenience wrapper for
\code{deSolve::ode()})}
\usage{
sim_rescomp(pars, stochastic = FALSE, totaltime, cinit, rinit, ...)
}
\arguments{
\item{pars}{S3 object of class \code{rescomp} returned by
\code{rescomp::spec_rescomp()}.}

\item{stochastic}{A boolean. If TRUE, use stochastic simulation,
via tau leaping instead of differential equations.}

\item{totaltime}{Numeric vector of length 1: the total simulation time.
If provided, overrides the value in \code{pars}.}

\item{cinit}{Numeric vector of length 1 or length \code{spnum} specifying
initial consumer state values (densities).
If provided, overrides the value in \code{pars}.}

\item{rinit}{Numeric vector of length 1 or length \code{resnum} specifying
initial resource state values (concentrations).
If provided, overrides the value in \code{pars}.}

\item{...}{Other arguments passed to \code{deSolve::ode()}}
}
\value{
A list of two comprising i) the model dynamics and ii) model
specifications.
}
\description{
Simulate resource competition (a convenience wrapper for
\code{deSolve::ode()})
}
\examples{
pars <- spec_rescomp()
results1 <- sim_rescomp(pars = pars)
plot_rescomp(results1)

results2 <- sim_rescomp(pars = pars, totaltime = 100, cinit = 1000)
plot_rescomp(results2)
}
