% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{spec_event_res_custom}
\alias{spec_event_res_custom}
\title{Define an event which instantaneously changes resource concentrations using an arbitrary function}
\usage{
spec_event_res_custom(func, spnum = NULL, resnum = NULL)
}
\arguments{
\item{func}{A function that takes \code{resources} (a numeric vector of resource concentrations) and \code{params} (a list of parameters) and returns a numeric vector of resource concentrations.}

\item{resnum}{The number of resources.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object suitable for building event schedules to pass \code{spec_rescomp()}.
}
\details{
If \code{resnum} is NULL, \code{spec_rescomp()} will attempt to infer it.
This is fine if the result is used directly in an event schedule, but may fail if the result must be combined with other \code{rescomp_event}s first.
}
\examples{
# TODO
}
