% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{spec_funcresp_custom}
\alias{spec_funcresp_custom}
\title{Define a functional response using an arbitrary function}
\usage{
spec_funcresp_custom(func, spnum = NULL, resnum = NULL)
}
\arguments{
\item{func}{A function that takes \code{resources} (a numeric vector of resource concentrations) and \code{params} (a list of parameters) and returns a matrix of growth rates of each species on each resource.}

\item{spnum}{The number of species; the number of rows in the matrix returned by \code{func}.}

\item{resnum}{The number of resources; the expected length of the \code{resources} argument to \code{func} and the number of columns in the matrix returned by \code{func}.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
}
\details{
If \code{spnum} or \code{resnum} are NULL, \code{spec_rescomp()} will attempt to infer them.
This is fine if the result is passed directly to \code{spec_rescomp()}, but may fail if the result must be combined with other \code{rescomp_funcresp}s first.
}
\examples{
# Type 1 functional response with fixed growth rates
spec_funcresp_custom(
  function(resources, params) {
    growth_rates <- c(0.2, 0.3)
    outer(growth_rates, resources)
  },
  spnum = 2
)
}
