% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ressupply.R
\name{spec_ressupply_custom}
\alias{spec_ressupply_custom}
\title{Create a resource supply rate using an arbitrary function}
\usage{
spec_ressupply_custom(func, resnum = NULL)
}
\arguments{
\item{func}{A function that takes \code{resources} (a numeric vector of resource concentrations) and \code{params} (a list of parameters) and returns a vector of rates of change of each resource.}

\item{resnum}{The number of resources; the expected length of the \code{resources} argument to \code{func} and the length of the vector returned by \code{func}.}
}
\value{
S3 object of class \code{rescomp_ressupply}.
}
\description{
Produces an object suitable to pass as the \code{ressupply} to \code{spec_rescomp()}.
}
\details{
If \code{resnum} is NULL, \code{spec_rescomp()} will attempt to infer it.
This is fine if the result is passed directly to \code{spec_rescomp()}, but may fail if the result must be combined with other \code{rescomp_ressupply} first.
}
\examples{
# Two resources, A and B, with constant supply of A, and A spontaneously converting to B
ressupply <- spec_ressupply_custom(
  function(resources, params) {
    conversion <- params$conversion * resources[1]
    return(c(params$supply - conversion, conversion))
  },
  resnum = 2
)
get_ressupply(ressupply, c(10, 20), list(supply = 3, conversion = 0.2))
}
