% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ressupply.R
\name{spec_ressupply_logistic}
\alias{spec_ressupply_logistic}
\title{Create a resource supply rate using logistic resource growth}
\usage{
spec_ressupply_logistic(r, k)
}
\arguments{
\item{r}{A vector or \code{rescomp_coefs_vector}, with one number per resource. The intrinsic growth rate of each resource.}

\item{k}{A vector or \code{rescomp_coefs_vector}, with one number per resource. The carrying capacity of each resource.}
}
\value{
S3 object of class \code{rescomp_ressupply}.
}
\description{
Produces an object suitable to pass as the \code{ressupply} to \code{spec_rescomp()}.
}
\examples{
ressupply <- spec_ressupply_logistic(r = rescomp_coefs_lerp(c(0.2, 0.3), c(0, 0), "growth_inhibition"), k = c(10, 20))
get_ressupply(ressupply, c(2, 10), list(growth_inhibition = 0))
get_ressupply(ressupply, c(20, 0), list(growth_inhibition = 0))
get_ressupply(ressupply, c(2, 10), list(growth_inhibition = 0.8))
}
