% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{time_vals}
\alias{time_vals}
\title{Timings of happenings}
\usage{
time_vals(total = 1000, step = 0.1, doround = TRUE, pulse)
}
\arguments{
\item{total}{total time}

\item{step}{step size}

\item{doround}{round time vector?}

\item{pulse}{pulsing interval}
}
\description{
Timings of happenings
}
