% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{apply_event}
\alias{apply_event}
\title{Applies a \code{rescomp_event} object to modify state variables}
\usage{
apply_event(event_obj, species, resources, params)
}
\arguments{
\item{event_obj}{An object of class \code{rescomp_event}.}

\item{species}{A vector of species concentrations.}

\item{resources}{A vector of resource concentrations.}

\item{params}{A list of time-dependent parameters.}
}
\value{
A vector of state variables, species concentrations followed by resource concentrations.
}
\description{
Applies the instantaneous changes specified by a \code{rescomp_event} object to the full set of state variables (species and resources).
}
\details{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_event} objects.
}
\examples{
# TODO
}
