% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{def_cr_ode}
\alias{def_cr_ode}
\title{Define consumer resource ODE function}
\usage{
def_cr_ode(t, y, pars)
}
\arguments{
\item{t}{Time to simulate over}

\item{y}{Vector of initial states}

\item{pars}{A list of model parameters}
}
\value{
Model formula to pass to \code{sim_rescomp()}
}
\description{
Define consumer resource ODE function
}
\examples{
# TODO
}
