% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{def_cr_ode}
\alias{def_cr_ode}
\title{Define consumer resource ODE function}
\usage{
def_cr_ode(Time, State, Pars, essential = FALSE, chemo = TRUE)
}
\arguments{
\item{Time}{time to simulate over}

\item{State}{vector of initial states}

\item{Pars}{a list}

\item{essential}{(default = FALSE)}

\item{chemo}{(default = TRUE)}
}
\value{

}
\description{
Define consumer resource ODE function
}
