% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{def_cr_ode}
\alias{def_cr_ode}
\title{Define consumer resource ODE function}
\usage{
def_cr_ode(Time, State, Pars)
}
\arguments{
\item{Time}{Time to simulate over}

\item{State}{Vector of initial states}

\item{Pars}{A list of model parameters}
}
\value{
Model formula to pass to \code{sim_rescomp()}
}
\description{
Define consumer resource ODE function
}
