% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_funcresp.R
\name{df_funcresp}
\alias{df_funcresp}
\title{Ge#' Generate ggplot friendly data frame for plotting functional responses}
\usage{
df_funcresp(pars, maxx, madj = FALSE)
}
\arguments{
\item{pars}{Parameter list from spec_rescomp()}

\item{maxx}{Maximum resource value to get percapita growth rates across
(for plotting)}

\item{madj}{Logical. Standardize by mortality. Default = FALSE.}
}
\value{
data frame
}
\description{
Ge#' Generate ggplot friendly data frame for plotting functional responses
}
\examples{
pars <- spec_rescomp()
df_funcresp(pars)

}
