% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event.R
\name{event_sp_mult}
\alias{event_sp_mult}
\title{Define an event in which the consumer populations are multiplied by some factor}
\usage{
event_sp_mult(species_mult)
}
\arguments{
\item{species_mult}{A numeric vector or \code{rescomp_coefs_vector} by which to multiply the current species concentrations. Should be greater than 1 for an increase, or less than 1 for a decrease. Should not be negative.}
}
\value{
S3 object of class \code{rescomp_event}.
}
\description{
Produces an event object a pulse of multiplicatively increased/decreased consumer populations, e.g. a density-independent disturbance which kills consumer species.
}
\examples{
# TODO
}
