% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_and_name.R
\name{frame_and_name}
\alias{frame_and_name}
\title{Convert ode object to dataframe and name columns}
\usage{
frame_and_name(odeobject, pars)
}
\arguments{
\item{odeobject}{ode object from}

\item{pars}{Model parameters from make_par_list()}
}
\value{
data frame
}
\description{
Convert ode object to dataframe and name columns
}
