% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcresp.R
\name{funcresp_type1}
\alias{funcresp_type1}
\title{Define a linear functional response}
\usage{
funcresp_type1(a)
}
\arguments{
\item{a}{A matrix or \code{rescomp_coefs_matrix}, with one row per species and one column per resource. The attack rate of each species on each resource.}
}
\value{
S3 object of class \code{rescomp_funcresp}.
}
\description{
Produces an object suitable to pass as the \code{funcresp} to \code{spec_rescomp()}.
Creates a linear or type 1 functional response with attack rate \code{a}.
mu_ij(R_j) = a_ij * R_j
}
\examples{
funcresp1 <- funcresp_type1(
  matrix(c(0.2, 0.4, 0.3, 0.2), nrow = 2)
)
get_funcresp(funcresp1, 2, c(10, 20), list())

funcresp2 <- funcresp_type1(
  rescomp_coefs_lerp(
    matrix(c(0.2, 0.4, 0.3, 0.2), nrow = 2),
    matrix(c(0.2, 0.1, 0.3, 0.1), nrow = 2),
    "antibiotic_concentration"
  )
)
get_funcresp(funcresp2, 2, c(10, 20), list(antibiotic_concentration = 0.5))
}
