% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_schedule.R
\name{get_event_times}
\alias{get_event_times}
\title{Gets the list of times from a \code{rescomp_event_schedule} object}
\usage{
get_event_times(event_schedule_obj, totaltime)
}
\arguments{
\item{event_schedule_obj}{An object of class \code{rescomp_event_schedule}.}

\item{totaltime}{The total time which the simulation will run for.}
}
\value{
A vector of times.
}
\description{
This function is normally only for internal use, but is exported to aid users in debugging their created \code{rescomp_event} objects.
}
\examples{
# TODO
}
