% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_state.R
\name{initiate_state}
\alias{initiate_state}
\title{Providing starting values for state variables}
\usage{
initiate_state(constart = 10, restart = vars$resconc, vars)
}
\arguments{
\item{constart}{Starting abundance for consumers}

\item{restart}{Starting concentration for resources}

\item{vars}{Model parameters from make_par_list()}
}
\value{
vector
}
\description{
Providing starting values for state variables
}
\examples{
initiate_state(vars = make_par_list())
}
