% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_list.R
\name{make_par_list}
\alias{make_par_list}
\title{Generate list of parameters for consumer-resource model}
\usage{
make_par_list(
  spnum = 2,
  resnum = 1,
  mumatrix,
  kmatrix,
  qmatrix,
  linear = TRUE,
  essential = FALSE,
  chemo = FALSE,
  mort = 0.03,
  resspeed = 1,
  resconc = 1,
  respulse = 0,
  timepars = FALSE,
  timeswitch = 96,
  timeswitch_length = 1000
)
}
\arguments{
\item{spnum}{number of consumers}

\item{resnum}{number of resources}

\item{mumatrix}{matrix of maximum growth rates (type 2)/ constant of proportionalities (type 1). # rows = spnum; # cols = resnum}

\item{kmatrix}{matrix of half saturation constants (type 2, ignored if linear = TRUE). # rows = spnum; # cols = resnum}

\item{qmatrix}{matrix of resource quotas. # rows = spnum; # cols = resnum}

\item{linear}{if FALSE equals type 2 function (i.e. Monod)}

\item{essential}{if FALSE resources are substitutable}

\item{chemo}{if FALSE resources grow logistically}

\item{mort}{density independent mortality rate}

\item{resspeed}{resource intrinsic rate of increase (if chemo = FALSE), otherwise chemostat dilution rate. Set to zero for pulsing only.}

\item{resconc}{resource carrying capacity (if chemo = FALSE), otherwise chemostat supply concentration}

\item{respulse}{resource pulse size. Requires events argument in call to \code{ode}.}

\item{timepars}{Logical. If time dependent parameters required.}

\item{timeswitch}{Integer. Frequency of parameter switching.}

\item{timeswitch_length}{Integer. Total time for parameter switching. Should be equivalent to total simulation time.}
}
\value{
list
}
\description{
Generate list of parameters for consumer-resource model
}
\examples{
make_par_list()
}
