% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crsim.R
\name{plot_crsim}
\alias{plot_crsim}
\title{Plot consumer and resource dynamics from deSolve output}
\usage{
plot_crsim(odeobj, pars)
}
\arguments{
\item{odeobj}{odeobject}

\item{pars}{parameters from make_par_list()}
}
\value{

}
\description{
Plot consumer and resource dynamics from deSolve output
}
