% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_crsim.R
\name{plot_crsim}
\alias{plot_crsim}
\title{Plot consumer and resource dynamics from deSolve output}
\usage{
plot_crsim(odeobj, pars)
}
\arguments{
\item{odeobj}{odeobject}

\item{pars}{parameters from make_par_list()}
}
\value{
ggplot object
}
\description{
Plot consumer and resource dynamics from deSolve output
}
\examples{

library(deSolve)
pars <- make_par_list()
happenings <- time_vals()
m1 <- ode(
    func = def_cr_ode,
    y = initiate_state(vars = pars),
    parms = pars,
    times = happenings$totaltime,
    method = "lsoda"
)
plot_crsim(m1, pars)

}
