% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcresp.R
\name{plot_funcresp}
\alias{plot_funcresp}
\title{Plot functional responses}
\usage{
plot_funcresp(pars, maxx)
}
\arguments{
\item{pars}{Parameter list from spec_rescomp()}

\item{maxx}{Resource value to calculate per-capita growth rates up to (xlim).}
}
\value{
ggplot object
}
\description{
Plot functional responses
}
