% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcresp.R
\name{plot_funcresp}
\alias{plot_funcresp}
\title{Plot functional responses}
\usage{
plot_funcresp(pars, maxx, madj = FALSE)
}
\arguments{
\item{pars}{Parameter list from spec_rescomp()}

\item{maxx}{Numeric vector of length 1.
Resource value to calculate per-capita growth rates up to (xlim).}

\item{madj}{Logical vector of length 1. Whether to standardize per capita
growth rates by mortality.}
}
\value{
ggplot object
}
\description{
Plot functional responses
}
\examples{

pars <- spec_rescomp()
plot_funcresp(pars)

pars <- spec_rescomp(
    spnum = 2,
    resnum = 2,
    funcresp = "type2",
    mumatrix = list(matrix(c(0.7,0.3,
                             0.4,0.5),
                           nrow = 2,
                           ncol = 2,
                           byrow = TRUE))
)
plot_funcresp(pars)
plot_funcresp(pars, madj = TRUE)

}
