% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcresp.R
\name{plot_funcresp}
\alias{plot_funcresp}
\title{Plot functional responses}
\usage{
plot_funcresp(pars, maxx)
}
\arguments{
\item{pars}{parameter list from make_par_list()}

\item{maxx}{maximum resource value to get percapita growth rates for (for plotting)}
}
\value{

}
\description{
Plot functional responses
}
