% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rescomp.R
\name{plot_rescomp}
\alias{plot_rescomp}
\title{Plot consumer and resource dynamics from deSolve output}
\usage{
plot_rescomp(
  model,
  consumers = TRUE,
  resources = TRUE,
  logy = FALSE,
  lwd = 1,
  xlims = NULL
)
}
\arguments{
\item{model}{List output from \code{sim_rescomp()}. First element is an object of
class deSolve. Second element is an object of class rescomp.}

\item{consumers}{Plot consumer dynamics? Default = TRUE.}

\item{resources}{Plot resource dynamics? Default = TRUE.}

\item{logy}{Log transfrom y-axis (default = FALSE).}

\item{lwd}{Line width (default = 1)}

\item{xlims}{Vector of length giving the time frame to plot.}
}
\value{
ggplot object
}
\description{
Plot consumer and resource dynamics from deSolve output
}
\examples{
pars <- spec_rescomp()
m1 <- sim_rescomp(pars)
plot_rescomp(m1)
}
