% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_rescomp.R
\name{print.rescomp}
\alias{print.rescomp}
\title{S3 print method for class rescomp}
\usage{
\method{print}{rescomp}(x, ..., detail = "summary")
}
\arguments{
\item{x}{rescomp object}

\item{...}{Further arguments}

\item{detail}{Either "summary" or "list". "summary" (default) provides plain language
summary info. "list" prints the full parameter list.}
}
\description{
S3 print method for class rescomp
}
