% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{rescomp_param_custom}
\alias{rescomp_param_custom}
\title{Create a rescomp parameter using an arbitrary function}
\usage{
rescomp_param_custom(func)
}
\arguments{
\item{func}{A function that takes \code{t} (time) and returns a number to use as a parameter.}
}
\value{
S3 object of class \code{rescomp_param}.
}
\description{
Produces an object suitable to include in a \code{rescomp_param_list}.
}
\examples{
# Repeated pulsing and exponential decay
antibiotic_conc <- rescomp_param_custom(function(t) {
  0.5^(10 * (t \%\% 1))
})
get_params(antibiotic_conc, 0)
get_params(antibiotic_conc, 0.5)
get_params(antibiotic_conc, 1)
}
