% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{rescomp_param_sine}
\alias{rescomp_param_sine}
\title{Create a rescomp parameter using a sine wave}
\usage{
rescomp_param_sine(period = 1, min = 0, max = 1, offset = 0)
}
\arguments{
\item{period}{The period of the sine wave.}

\item{min}{The minimum value of the parameter; the mean minus the amplitude.}

\item{max}{The maximum value of the parameter; the mean plus the amplitude.}

\item{offset}{The phase shift of the sine wave; the time at which its value is equal to the mean.}
}
\value{
S3 object of class \code{rescomp_param}.
}
\description{
Produces an object suitable to include in a \code{rescomp_param_list}.
}
\examples{
# TODO
}
