% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rescomp.R
\name{sim_rescomp}
\alias{sim_rescomp}
\title{Simulate resource competition (a convenience wrapper for
\code{deSolve::ode})}
\usage{
sim_rescomp(parms, times, y = initiate_state(parms), events = NULL, ...)
}
\arguments{
\item{parms}{Parameter list returned by \code{rescomp::make_par_list}.}

\item{times}{List. Total simulation time and pulsing if relevant
(use \code{rescomp::time_vals})}

\item{y}{Initial values for state variables.}

\item{events}{NULL}

\item{...}{Other arguments passed to \code{deSolve::ode}}
}
\value{
Matrix of class deSolve
}
\description{
Simulate resource competition (a convenience wrapper for
\code{deSolve::ode})
}
\examples{
pars <- make_par_list()
happenings <- time_vals(total = 1000)
init <- initiate_state(pars)
sim_rescomp(y = init,
    parms = pars,
    times = happenings)

}
