% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rescomp.R
\name{sim_rescomp}
\alias{sim_rescomp}
\title{Simulate resource competition (a convenience wrapper for
\code{deSolve::ode})}
\usage{
sim_rescomp(parms, times, y, events = NULL, ...)
}
\arguments{
\item{parms}{Parameter list returned by \code{rescomp::make_par_list}.}

\item{times}{List of up to 2 giving total simulation time and pulsing
sequence where relevant (use \code{rescomp::time_vals}). If provided,
overrides values given in parms.}

\item{y}{Vector of initial values for state variables. If provided,
overrides values given in parms.}

\item{events}{NULL (events specified in \code{make_par_list})}

\item{...}{Other arguments passed to \code{deSolve::ode}}
}
\value{
Matrix of class deSolve
}
\description{
Simulate resource competition (a convenience wrapper for
\code{deSolve::ode})
}
\examples{
pars <- make_par_list()
happenings <- time_vals(total = 1000)
init <- initiate_state(pars)
sim_rescomp(y = init,
    parms = pars,
    times = happenings)

}
