% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_rescomp.R
\name{spec_rescomp}
\alias{spec_rescomp}
\title{Generate list of parameters for a consumer-resource model to be
passed to \code{sim_rescomp()}}
\usage{
spec_rescomp(
  spnum = 1,
  resnum = 1,
  funcresp = funcresp_type1(matrix(0.1, nrow = spnum, ncol = resnum)),
  quota = matrix(0.001, nrow = spnum, ncol = resnum),
  essential = FALSE,
  mort = 0.03,
  ressupply = ressupply_chemostat(0.03, 1),
  params = rescomp_param_list(),
  events = list(),
  totaltime = 1000,
  cinit = 10,
  rinit = 1,
  verbose = TRUE
)
}
\arguments{
\item{spnum}{Integer vector of length 1: the number of consumers.}

\item{resnum}{Integer vector of length 1: the number of resources.}

\item{funcresp}{An object of class \code{rescomp_funcresp} specifying the functional response.}

\item{quota}{Numeric matrix or \code{rescomp_coefs_matrix}, the elements of which
give the resource quotas. The number of rows and columns should be equal to
\code{spnum} and \code{resnum} respectively.}

\item{essential}{Logical vector of length 1. If FALSE resources are substitutable.}

\item{mort}{Numeric vector or \code{rescomp_coefs_vector} of length \code{spnum},
specifying density independent mortality rates.}

\item{ressupply}{An object of class \code{rescomp_ressupply} specifying the resource supply.}

\item{params}{An object of class \code{rescomp_param_list} specifying a set of parameters which vary
with time, on which other parameters of the model (e.g. funcresp, ressupply) may depend.}

\item{events}{A list of objects of class \code{rescomp_event_schedule}, specifying events that
instantaneously change consumer or resource densities.}

\item{totaltime}{Numeric vector of length 1: the total simulation time.}

\item{cinit}{Numeric vector of length 1 or length \code{spnum} specifying
initial consumer state values (densities).}

\item{rinit}{Numeric vector of length 1 or length \code{resnum} specifying
initial resource state values (concentrations).}

\item{verbose}{If TRUE (default) prints model and simulation summary specs.}
}
\value{
S3 object of class \code{rescomp}.
}
\description{
Generate list of parameters for a consumer-resource model to be
passed to \code{sim_rescomp()}
}
\examples{
# TODO
}
