% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_rescomp.R
\name{spec_rescomp}
\alias{spec_rescomp}
\title{Generate list of parameters for a consumer-resource model to be
passed to \code{sim_rescomp()}}
\usage{
spec_rescomp(
  spnum = 1,
  resnum = 1,
  funcresp = funcresp_type1(crmatrix(0.1)),
  quota = crmatrix(0.001),
  efficiency,
  essential = FALSE,
  mort = 0.03,
  ressupply = ressupply_chemostat(0.03, 1),
  params = rescomp_param_list(),
  events = list(),
  extra_terms = list(),
  totaltime = 1000,
  cinit = 10,
  rinit = 1,
  verbose = FALSE
)
}
\arguments{
\item{spnum}{Integer vector of length 1: the number of consumers.}

\item{resnum}{Integer vector of length 1: the number of resources.}

\item{funcresp}{An object of class \code{rescomp_funcresp} specifying the functional response.}

\item{quota}{Numeric matrix or \code{rescomp_coefs_matrix}, the elements of which
give the resource quotas. The number of rows and columns should be equal to
\code{spnum} and \code{resnum} respectively.
Mutually exclusive with \code{efficiency}.}

\item{efficiency}{Numeric matrix or \code{rescomp_coefs_matrix}, the elements of which
give the efficiency of each consumer on each resource. The number of rows
and columns should be equal to \code{spnum} and \code{resnum} respectively.
Mutually exclusive with \code{quota}.}

\item{essential}{Logical vector of length 1. If FALSE resources are substitutable.}

\item{mort}{Numeric vector or \code{rescomp_coefs_vector} of length \code{spnum},
specifying density independent mortality rates.}

\item{ressupply}{An object of class \code{rescomp_ressupply} specifying the resource supply.}

\item{params}{An object of class \code{rescomp_param_list} specifying a set of parameters which vary
with time, on which other parameters of the model (e.g. funcresp, ressupply) may depend.}

\item{events}{A list of objects of class \code{rescomp_event_schedule}, specifying events that
instantaneously change consumer or resource densities.}

\item{extra_terms}{A list of lists. Each list details a set of extra terms
that should be grouped as a single transition if using stochastic simulation.
Each inner list must three arguments: \code{consumers}, \code{resources}, and \code{rate}.
\code{consumers} and \code{resources} are vectors of length \code{spnum} and \code{resnum} respectively,
specifying the size of the change to the corresponding variables from the transition.
\code{rate} is a function returning the rate of the transition.
Rate functions take three arguments: a vector of consumer concentrations,
a vector of resource concentrations, a vector of total growth rates of each consumer.}

\item{totaltime}{Numeric vector of length 1: the total simulation time.}

\item{cinit}{Numeric vector of length 1 or length \code{spnum} specifying
initial consumer state values (densities).}

\item{rinit}{Numeric vector of length 1 or length \code{resnum} specifying
initial resource state values (concentrations).}

\item{verbose}{If TRUE (default) prints model and simulation summary specs.}
}
\value{
S3 object of class \code{rescomp}.
}
\description{
Generate list of parameters for a consumer-resource model to be
passed to \code{sim_rescomp()}
}
\details{
Only one of \code{efficiency} and \code{quota} should be specified. Specifying both is an error.
The default, if neither is specified, is to use \code{quota}.
If using \code{quota}, the functional responses are taken to give per capita growth rates.
If using \code{efficiency}, the functional responses are taken to give attack rates.
}
\examples{
m1 <- spec_rescomp()
plot_rescomp(sim_rescomp(m1))

m2 <- spec_rescomp(
  spnum = 2,
  resnum = 2,
  funcresp = funcresp_type1(
    a = crmatrix(
      0.1, 0.2,
      0.15, 0.15
    )
  )
)
plot_rescomp(sim_rescomp(m2))

m3 <- spec_rescomp(
  ressupply = ressupply_constant(0),
  events = list(
    event_schedule_periodic(
      event_batch_transfer(dilution = 0.1, resources = 1),
      period = 250
    )
  )
)
plot_rescomp(sim_rescomp(m3))
}
