% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_cr_ode.R
\name{time_vals}
\alias{time_vals}
\title{Timings of happenings}
\usage{
time_vals(total = 1000, step = 0.1, doround = TRUE, pulse)
}
\arguments{
\item{total}{Total time.}

\item{step}{Step size.}

\item{doround}{Round time units (handles issues with numerical differences
that produce warning messages when pulsing resources and/or consumers).}

\item{pulse}{Pulsing interval.}
}
\description{
Timings of happenings
}
\examples{

time_vals(1000, pulse = 100)
}
